+++ src/rendering/dashboard.cpp
@@ -62,17 +62,17 @@ void Dashboard::addDashboardItem(std::unique_ptr<DashboardItem> item) {
 }
 
 void Dashboard::removeDashboardItem(int index) {
-    ghoul_assert(index < _items.size(), "Invalid index");
+    ghoul_assert(index < static_cast<int>(_items.size()), "Invalid index");
     removePropertySubOwner(_items[index].get());
     _items.erase(_items.begin() + index);
 }
 
 bool Dashboard::hasItem(int index) const {
-    return (index >= 0) && (index < _items.size());
+    return (index >= 0) && (index < static_cast<int>(_items.size()));
 }
 
 const DashboardItem& Dashboard::item(int index) const {
-    ghoul_assert(index < _items.size(), "Invalid index");
+    ghoul_assert(index < static_cast<int>(_items.size()), "Invalid index");
     return *_items[index];
 }
